frontend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.0.0
Flask-Mail==0.9.1
itsdangerous==2.1.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Werkzeug==3.0.1
